Initial commit
Imported all functionality, tests, and documentation from zend-mvc. Updated
tests to test the plugin directly (instead of via the controller), and split off
tests that needed to mock/intercept the file upload operations so that they can
run in a separate process. Added a Module class for seeding the controller
plugin manager.