qa: allow mocking move_uploaded_file, is_uploaded_file when testing
Prior to laminas-validator 2.14.5 and laminas-filter 2.13.0, these libraries did not import global functions.
After those versions, however, we did (due to our CS), which meant that our ability to mock the `is_uploaded_file()` and `move_uploaded_file()` functions _in those namespaces_ no longer had an effect.

This patch adds a test bootstrap that copies the `Laminas\Validator\File\UploadFile` class and `Laminas\Filter\File\RenameUpload` class to a polyfill directory, and comments out the import of the above listed functions.
This approach allows us to mock how they behave when testing.

The files are NOT checked into the repository, AND are recreated each time the boostrap is run (to ensure we have the version installed to test against).

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>